adds favico and moves the gtag to meta
