+++ .gitignore
@@ -1,2 +1,2 @@
- .DS_Store
+.DS_Store
 node_modules/
